Added a "poorly improved" quality check in group analysis
Group CICADA tries to determine data outliers that may need to be manually IC adjusted or removed. This is labeled as a "CICADA outlier". In this edit, a qualifier for data to be a CICADA outlier is added. This qualifier is called "poorly improved data". Data is marked as poorly improved (and thus an automatic CICADA outlier) if, when comparing before and after cleaning any of the following are not improved: GM overlap, Smoothing retention, best hrf power overlap, FD corr, DVARS corr, or temporal spikiness. This is also reflected in the group qc table.