Create more useful Group IC output
A lighter version of CICADA is performed on the Group ICs. Outputs information on smoothness, gray matter overlap, and powerspectra overlap of dual regression group ICs. Also calculates a neural signal probability in the same manner as Auto CICADA from these values. The signal labels are not meant to be 100% accurate (as Group ICA results will follow different patterns than at the subject level). However, this is meant to help with guidance when looking at Group level ICs.