Integrated despiking option
Integrated option to despike fMRI data before IC decomposition. In theory, this may help MELODIC IC better separate out signal from noise, which may lead to improved CICADA classification (at least potentially for data with large movement spikes). However, this has not been thoroughly tested and was not a part of the original manuscript. As such, default is to not despike.